+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kqueiroz <kqueiroz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/09 13:56:57 by kqueiroz          #+#    #+#             */
-/*   Updated: 2025/08/24 20:51:05 by kqueiroz         ###   ########.fr       */
+/*   Updated: 2025/09/02 21:13:39 by kqueiroz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	ft_putchar_fd(char c, int fd);
 int	ft_putstr_fd(char *s, int fd);
 int	ft_putnbr_fd(int n, int fd);
 int	ft_putunsigned_fd(unsigned int n, int fd);
-int	ft_puthex_fd(unsigned long n, char format, int fd);
+int	ft_puthex_fd(unsigned int n, char format, int fd);
 int	ft_putptr_fd(void *ptr, int fd);
 
 #endif
