+++ ft_puthex_fd.c
@@ -6,16 +6,16 @@
 /*   By: kqueiroz <kqueiroz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/09 20:36:51 by kqueiroz          #+#    #+#             */
-/*   Updated: 2025/08/24 20:46:40 by kqueiroz         ###   ########.fr       */
+/*   Updated: 2025/09/02 20:46:24 by kqueiroz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthex_fd(unsigned long n, char format, int fd)
+int	ft_puthex_fd(unsigned int n, char format, int fd)
 {
-	int		count;
-	char	*base;
+	unsigned int	count;
+	char			*base;
 
 	count = 0;
 	if (fd < 0)
